pyproject.toml
[project]
name = "boa-favorites-cu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv",
    "titanoboa>=0.2.5",
    "web3>=7.6.0",
]
